+++ js/gen/app.config.js
@@ -1,21 +1,20 @@
 app.config = {
 	env					: 'prod',
-	debug 			: true,
-	version				: '1.8.49', // Install Python button added to Advanced Table Settings, Demo pattern added to Advanced Lights, Warm white balance adjustment & Auto -resume previous pattern after leaving calibrate lights page.
+	version				: '1.8.62', // multiple download added to communtiy, added community model.
 	disconnect_timeout_to_stop_polling: 45000, // stop trying to find tables after 45 seconds
 	envs	: {
 		alpha: {	// loads local data only **5 tap on No Table Found Screen to launch shell app
 			base_url	: 'http://app.dev.withease.io:3001/', // local
-			api_url		: 'https://api.sisyphus.withease.io/',
-			web_url		: 'http://dev.webcenter.sisyphus-industries.com/',
+			api_url		: 'https://webcenter.sisyphus-industries.com/',
+			web_url		: 'https://webcenter.sisyphus-industries.com/',
 			sisbot_url  : 'http://api.dev.withease.io:3000/', //talking to sisbot
 			port		: 3001,
 		},
 		beta: {		// tests local network
 			base_url	: 'http://app.dev.withease.io:3001/', //local url
-			api_url		: 'https://webcenter.sisyphus-industries.com/', // add entry in your computers /etc/hosts mapped to your bot's IP address
-			web_url		: 'https://webcenter.sisyphus-industries.com/', //web_center url	***Change to this for Rails web_center= http://localhost:3000/  (aka rails s) //  10.0.0.3	beta_bot.local
-			sisbot_url  : 'http://192.168.86.20:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
+			api_url		: 'http://localhost:3000/', // add entry in your computers /etc/hosts mapped to your bot's IP address
+			web_url		: 'http://localhost:3000/', //web_center url	***Change to this for Rails web_center= http://localhost:3000/  (aka rails s) //  10.0.0.3	beta_bot.local
+			sisbot_url  : 'http://192.168.1.6:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
 			port		: 3001,
 		},
 		sisbot: {