+++ js/models/app.model.sisbot.js
@@ -39,6 +39,8 @@ app.model.sisbot = {
 				sisbot	: '-',
 			},
 
+			csons	: 'false', // available cson files
+
 			has_software_update				: 'false',
 			is_connected					: false,
 			is_socket_connected				: 'false',
@@ -58,8 +60,10 @@ app.model.sisbot = {
 			default_settings				: {},
 			default_settings_error			: 'false',
 
-			log_date						: moment().format('MM/DD/YYYY'),
-			log_type						: 'sisbot',		// sisbot|plotter|proxy
+			// log_date						: moment().format('MM/DD/YYYY'),
+			// log_type						: 'sisbot',		// sisbot|plotter|proxy
+			log_file						: 'false', // currently selected for download
+			log_files						: [],
 			uploading_track			: 'false',
 
 			wait_for_send				: 'false', // don't send request before hearing response
@@ -67,6 +71,7 @@ app.model.sisbot = {
 			rem_pattern					: 'false',
 			rem_primary_color		: '0xFFFFFFFF',
 			rem_secondary_color	: '0x00FFFFFF',
+			show_picker					: 'true',
 
 			edit		: {},
 			data		: {
@@ -95,8 +100,8 @@ app.model.sisbot = {
 				is_internet_connected: 'false',
 				is_network_connected: 'false',
 				is_network_separate : 'false',
-        wifi_network        : '',
-        wifi_password       : '',
+				wifi_network        : '',
+				wifi_password       : '',
 				failed_to_connect_to_wifi: 'false',
 				wifi_forget			: 'false',
 
@@ -138,7 +143,7 @@ app.model.sisbot = {
 				table_settings: {}, // Advanced table settings, overrides CSON on reboot
 
 				led_enabled: 'false',
-				led_pattern_ids: ['white','solid','fade','spread','comet','rainbow'],
+				led_pattern_ids: ['white','solid','fade','spread','comet','rainbow','paint','demo'],
 				led_pattern: 'white',
 				led_offset : 0,
 				led_primary_color: '0xFFFFFFFF', // Hex
@@ -198,6 +203,11 @@ app.model.sisbot = {
 		if (this.get('data.sleep_time') == 'false') // fix is_sleep_enabled toggle
 			this.set('data.is_sleep_enabled', 'false');
 
+		if (this.get('data.led_pattern') != 'false') {
+			this.setup_edit();
+			this._update_pattern_colors();
+		}
+
 		this._poll_state();
 	},
 	update_network: function() {
@@ -245,7 +255,11 @@ app.model.sisbot = {
 		console.log("_update_sisbot()", endpoint, data);
 		if (!_timeout) _timeout = 5000;
 
-		if (app.config.env == 'alpha') return this;
+		if (app.config.env == 'alpha') {
+			console.log('ALPHA is_internet_connected ==', this.get('data.is_internet_connected'));
+			this.set('data.is_internet_connected', 'true'); //setting to true for Apple to test Community
+			return this;
+		}
 
 		var self	= this;
 		var address	= this.get('data.local_ip');
@@ -281,103 +295,7 @@ app.model.sisbot = {
 		}, 0);
 	},
 
-	// _fetching_cloud: false,
-	// _fetch_cloud: function () {
-	// 	console.log("_fetch_cloud()");
-	// 	if (this._fetching_cloud) 	return this;
-	//
-	// 	var self = this;
-	// 	this._fetching_cloud = true;
-	//
-	// 	var current_ip	= this.get('data.local_ip');
-	//
-	// 	app.post.fetch(exists = {
-	// 		_url	: app.config.get_api_url(),
-	// 		_type	: 'GET',
-	// 		_timeout: 1250,
-	// 		endpoint: 'sisbot_state/' + this.id,
-	// 	}, function exists_cb(obj) {
-	// 		self._fetching_cloud = false;
-	// 		// debugger;
-	// 		// console.log("_fetch_cloud() returned " + JSON.stringify(obj));
-	// 		if (obj.resp && obj.resp.local_ip) {
-	// 			// we are internet connected!
-	// 			var ip_address = obj.resp.local_ip;
-	// 			self.set('data.local_ip', ip_address);
-	// 		}
-	// 	}, 0);
-	// },
-	// _fetching_bluetooth: false,
-	// _fetch_bluetooth: function () {
-	// 	console.log("_fetch_bluetooth()");
-	// 	if (!app.is_app)				return this;
-	// 	if (this._fetching_bluetooth) 	return this;
-	//
-	// 	var self = this;
-	// 	this._fetching_bluetooth = true;
-	//
-	// 	var current_ip	= this.get('data.local_ip');
-	// 	var sub_id		= this.id.substr(this.id.length - 7);
-	//
-	// 	if (this.is_legacy()) sub_id = 'sisyphus';
-	//
-	// 	app.manager.start_ble_scan(sub_id, function (ip_address) {
-	// 		self._fetching_bluetooth = false;
-	//
-	// 		if (!ip_address) {
-	// 			// no ip address. must be doing network stuff
-	// 		} else if (current_ip == ip_address && ip_address == '192.168.42.1') {
-	// 			// sisyphus is in hotspot mode and we failed to connect to it
-	// 			self.set('data.reason_unavailable', 'connect_to_wifi');
-	// 		} else if (current_ip !== ip_address) {
-	// 			// we successfully connected to wifi!
-	// 			self.set('data.local_ip', ip_address);
-	// 		}
-	// 	});
-	// },
-	// _fetch_network: function () {
-	// 	console.log("_fetch_network()");
-	// 	if (!app.is_app) return this;
-	//
-	// 	var self = this;
-	//
-	// 	app.manager.get_network_ip_address(function(ip_address) {
-	// 		if (!ip_address)	return self;
-	//
-	// 		var ip_add	= ip_address.split('.');
-	// 		ip_add.pop();
-	//
-	// 		var ip_base = ip_add.join('.');
-	// 		var count = 0;
-	//
-	// 		_.each(_.range(0, 256), function(num) {
-	// 			self._ping_sisbot(ip_base + '.' + num);
-	// 		});
-	// 	});
-	//
-	// 	return this;
-	// },
-	// _ping_sisbot: function(hostname) {   // trying to find what sisbots are available.
-	// 	console.log("_ping_sisbot()", hostname);
-	// 	var self = this;
-	//
-	// 	app.post.fetch(exists = {
-	// 		_url	: 'http://' + hostname + '/',
-	// 		_type	: 'POST',
-	// 		_timeout: 1250,
-	// 		endpoint: 'sisbot/exists',
-	// 		data	: {}
-	// 	}, function exists_cb(obj) {
-	// 		if (!obj.resp || !obj.resp.hostname)
-	// 			return self;
-	//
-	// 		if (obj.resp.id == self.id) {
-	// 			app.manager.intake_data(obj.resp);
-	// 		}
-	// 	}, 0);
-	//
-	// 	return this;
-	// },
+
 	_check_serial: function () {
 		console.log("_check_serial()");
 
@@ -396,15 +314,15 @@ app.model.sisbot = {
 		*/
 	},
 	_update_timestamp: function() {
-		console.log("_update_timestamp()");
+		// console.log("_update_timestamp()");
 		this.set('timestamp', ''+Date.now());
-		console.log("Update Timestamp", ''+Date.now(), this.get('timestamp'));
+		// console.log("Update Timestamp", ''+Date.now(), this.get('timestamp'));
 	},
 	/**************************** sockets ********************************/
 	_socket_connect: function() {
 		var self = this;
 
-		console.log("Sisbot: Socket Connect");
+		// console.log("Sisbot: Socket Connect");
 
 		this.set('is_socket_connected', 'true');
 		this.set('is_polling', "false");
@@ -422,11 +340,11 @@ app.model.sisbot = {
 		// }, 10000);
 	},
 	_socket_disconnect: function() {
-		console.log("_socket_disconnect()");
+		// console.log("_socket_disconnect()");
 		this.set('is_socket_connected', 'false');
 
 		var self = this;
-		console.log("Sisbot: Socket Disconnect");
+		// console.log("Sisbot: Socket Disconnect");
 
 		if (this.get('is_polling') == "false") {
 			setTimeout(function() {
@@ -436,7 +354,7 @@ app.model.sisbot = {
 		}
 	},
 	_socket_error: function(data) {
-		console.log("Sisbot: Socket Error", data);
+		// console.log("Sisbot: Socket Error", data);
 		if (this.get('is_polling') == "false") {
 			this.set('is_polling', "true");
 			this._poll_state();
@@ -445,7 +363,7 @@ app.model.sisbot = {
 	/**************************** POLLING *************************************/
 	_poll_timer: false,
 	_poll_failure: function () {
-		console.log("_poll_failure()");
+		// console.log("_poll_failure()");
 		if (this._poll_timer == false) {
 			this._poll_timer = moment();
 			this._retry_find = true;
@@ -483,7 +401,7 @@ app.model.sisbot = {
 		return this;
 	},
 	_poll_failure_stop: function () {
-		console.log("_poll_failure_stop()");
+		// console.log("_poll_failure_stop()");
 		if (this._poll_then_reset_bool == true) {
 			window.location.reload();
 		}
@@ -492,24 +410,25 @@ app.model.sisbot = {
 				   .set('sisbot_reconnecting', 'false');
 	},
 	_poll_restart: function () {
-		console.log("_poll_restart()");
+		// console.log("_poll_restart()");
 		this._poll_timer = false;
 		this.set('is_polling', 'true');
 		this._poll_state();
 	},
 	_poll_then_reset_bool: false,
 	_poll_then_reset: function() {
-		console.log("_poll_then_reset()");
+		// console.log("_poll_then_reset()");
 		this._poll_then_reset_bool = true;
 		this._poll_restart();
 	},
 	_poll_state: function () {
-		console.log("_poll_state()");
+		// console.log("_poll_state()");
 		var self = this;
 
 		if (app.config.env == 'alpha') {
 			// FOR APPLE TESTING...
-			app.manager.set('is_sisbot_available', 'true')
+			app.manager.set('is_sisbot_available', 'true');
+			this.set('data.is_internet_conected', 'true');
 		}
 
 		if (this.get('is_master_branch') == 'false') console.log("Get State: ", app.manager.get('is_sisbot_available'), this.get('is_polling'));
@@ -1116,12 +1035,28 @@ app.model.sisbot = {
 			}
 		});
 	},
+	load_log_files: function() {
+		console.log("Load Log Files");
+		var self = this;
+
+		// load log files from sisbot
+		this._update_sisbot('get_log_filenames', {}, function(obj) {
+			if (obj.err) return console.log("Error loading log files", obj.err);
+
+			if (obj.resp && _.isArray(obj.resp)) {
+				self.set('log_files', obj.resp);
+				if (obj.resp.indexOf('proxy.log') >= 0) self.set('log_file', 'proxy.log');
+				else if (obj.resp.length > 0) self.set('log_file', obj.resp[0]);
+			}
+		});
+	},
 	get_log_file: function() {
 		// { filename: 'YYYYMMDD_sisbot|plotter|proxy', }
 
-		var date 		= this.get('log_date').split('/');
-		var type 		= this.get('log_type');
-		var file 		= date[2] + date[0] + date[1] + '_' + type;
+		// var date 		= this.get('log_date').split('/');
+		// var type 		= this.get('log_type');
+		// var file 		= date[2] + date[0] + date[1] + '_' + type;
+		var file 			= this.get('log_file');
 		var file_url 	= 'http://' + this.get('data.local_ip') + '/sisbot/download_log_file/' + file;
 
 		app.plugins.file_download(file_url);
@@ -1365,18 +1300,52 @@ app.model.sisbot = {
 	},
 	_change_led_pattern: function() {
 		var self = this;
-		// console.log("Change Pattern");
 
 		var new_pattern = this.get('edit.led_pattern');
+
 		if (new_pattern != 'false' && this.get('data.led_pattern') != new_pattern) {
 			this.set('data.led_pattern', new_pattern);
 
 			// send to sisbot
 			var pattern = app.collection.get(new_pattern);
-			console.log("Update Pattern", JSON.stringify(pattern.get('data')));
-			self._update_sisbot('set_led_pattern', pattern.get('data'), function (obj) {
-				// do nothing
-			});
+			if (pattern) {
+				console.log("Update LED Pattern", pattern.get('data'));
+				self._update_sisbot('set_led_pattern', pattern.get('data'), function (obj) {
+					if (obj.err) return console.error(err);
+
+					// fix possible incorrect return value
+					if (_.isObject(obj.resp.led_primary_color)) {
+						var colors = obj.resp.led_primary_color;
+						var red = colors.red.toString(16);
+						if (red.length < 2) red = '0'+red;
+						var green = colors.green.toString(16);
+						if (green.length < 2) green = '0'+green;
+						var blue = colors.blue.toString(16);
+						if (blue.length < 2) blue = '0'+blue;
+						var white = colors.white.toString(16);
+						if (white.length < 2) white = '0'+white;
+						obj.resp.led_primary_color = '0x'+red+green+blue+white;
+						console.log("Fix Primary", JSON.stringify(colors), obj.resp.led_primary_color);
+					}
+					if (_.isObject(obj.resp.led_secondary_color)) {
+						var colors = obj.resp.led_secondary_color;
+						var red = colors.red.toString(16);
+						if (red.length < 2) red = '0'+red;
+						var green = colors.green.toString(16);
+						if (green.length < 2) green = '0'+green;
+						var blue = colors.blue.toString(16);
+						if (blue.length < 2) blue = '0'+blue;
+						var white = colors.white.toString(16);
+						if (white.length < 2) white = '0'+white;
+						obj.resp.led_secondary_color = '0x'+red+green+blue+white;
+						console.log("Fix Secondary", JSON.stringify(colors), obj.resp.led_secondary_color);
+					}
+
+					// intake data
+					app.manager.intake_data(obj.resp);
+					console.log("Set LED Pattern return", obj);
+				});
+			}
 		}
 	},
 	_update_pattern_colors: function() {
@@ -1384,14 +1353,14 @@ app.model.sisbot = {
 
 		// load last used color from this pattern
 		var pattern = app.collection.get(this.get('data.led_pattern'));
-		console.log("Update Pattern Colors", pattern.get('data'));
+		// console.log("Update Pattern Colors", pattern.get('data'), this.get('edit.led_primary_color'), this.get('edit.led_secondary_color'));
 		if (pattern) {
 			if (pattern.get('data.is_white') == 'true') {
-				console.log("Update white from led_pattern", pattern.get('data'));
+				// console.log("Update white from led_pattern", pattern.get('data'));
 				this.set('edit.led_primary_color', pattern.get_white_color());
 				this.led_color();
 			} else if (pattern.get('data.is_primary_color') == 'true' || pattern.get('data.is_secondary_color') == 'true') {
-				console.log("Update colors from led_pattern", pattern.get('data'));
+				// console.log("Update colors from led_pattern", pattern.get('data'));
 				if (pattern.get('data.is_primary_color') == 'true' && pattern.get('data.led_primary_color') != 'false') this.set('edit.led_primary_color', pattern.get('data.led_primary_color'));
 				if (pattern.get('data.is_secondary_color') == 'true' && pattern.get('data.led_secondary_color') != 'false') this.set('edit.led_secondary_color', pattern.get('data.led_secondary_color'));
 				this.led_color();
@@ -1404,7 +1373,7 @@ app.model.sisbot = {
 		return do_save;
 	},
 	remember_colors: function(data) {
-		console.log("Remember LED_Colors", data);
+		console.log("Remember LED_Colors", this.get('data.led_primary_color'), this.get('data.led_secondary_color'));
 
 		this.set('rem_primary_color', this.get('data.led_primary_color'));
 		this.set('rem_secondary_color', this.get('data.led_secondary_color'));
@@ -1417,8 +1386,17 @@ app.model.sisbot = {
 
 		app.trigger('session:active', { 'secondary': 'false' });
 	},
+	toggle_picker: function(data) {
+		console.log("Toggle Picker", this.get('edit.led_primary_color'), this.get('edit.led_secondary_color'));
+
+		this.set('show_picker', 'false');
+
+		this.led_color(null);
+
+		this.set('show_picker', 'true');
+	},
 	led_color: function(data) {
-		console.log("Sisbot LED_Color", data, this.get('edit.led_primary_color'), this.get('edit.led_secondary_color'));
+		// console.log("Sisbot LED_Color", data, this.get('edit.led_primary_color'), this.get('edit.led_secondary_color'));
 		var self = this;
 		var color_data = {};
 
@@ -1426,14 +1404,23 @@ app.model.sisbot = {
 
 		// check for primary change
 		var edit_primary = this.get('edit.led_primary_color');
-		if (edit_primary.match(/^0x/)) edit_primary = edit_primary.replace(/^0x/, '#');
-		console.log("Compare Primary Color", this.get('data.led_primary_color'), edit_primary);
+
+		// fix possible errors
+		if (edit_primary) {
+			if (edit_primary.match(/^0x/)) edit_primary = edit_primary.replace(/^0x/, '#');
+			if (!edit_primary.match(/^#[0-9a-f]{8}$/i)) {
+				edit_primary = this.get('data.led_primary_color');
+				this.set('edit.led_primary_color', edit_primary);
+			}
+		}
+
+		// console.log("Compare Primary Color", this.get('data.led_primary_color'), edit_primary);
 		if (this.get('data.led_primary_color') != edit_primary) {
 			this.set('data.led_primary_color', edit_primary);
 
 			console.log("Update Primary Color", this.get('data.led_primary_color'));
 
-			color_data.primary_color = edit_primary;
+			color_data.led_primary_color = edit_primary;
 
 			// update the led_pattern
 			if (led_pattern) led_pattern.set('data.led_primary_color', edit_primary);
@@ -1441,14 +1428,23 @@ app.model.sisbot = {
 
 		// check for secondary change
 		var edit_secondary = this.get('edit.led_secondary_color');
-		if (edit_secondary.match(/^0x/)) edit_secondary = edit_secondary.replace(/^0x/, '#');
-		console.log("Compare Secondary Color", this.get('data.led_secondary_color'), edit_secondary);
+
+		// fix possible errors
+		if (edit_secondary) {
+			if (edit_secondary.match(/^0x/)) edit_secondary = edit_secondary.replace(/^0x/, '#');
+			if (!edit_secondary.match(/^#[0-9a-f]{8}$/i)) {
+				edit_secondary = this.get('data.led_secondary_color');
+				this.set('edit.led_secondary_color', edit_secondary);
+			}
+		}
+
+		// console.log("Compare Secondary Color", this.get('data.led_secondary_color'), edit_secondary);
 		if (this.get('data.led_secondary_color') != edit_secondary) {
 			this.set('data.led_secondary_color', edit_secondary);
 
 			console.log("Update Secondary Color", this.get('data.led_secondary_color'));
 
-			color_data.secondary_color = edit_secondary;
+			color_data.led_secondary_color = edit_secondary;
 
 			// update the led_pattern
 			if (led_pattern) led_pattern.set('data.led_secondary_color', edit_secondary);
@@ -1459,6 +1455,7 @@ app.model.sisbot = {
 			console.log("Save color data", this.get('data.led_primary_color'), this.get('data.led_secondary_color'));
 			var save_data = [this.get('data')];
 			if (led_pattern) save_data.push(led_pattern.get('data'));
+			console.log("Save data", save_data);
 			this.save_to_sisbot(save_data);
 
 			this._update_sisbot('set_led_color', color_data, function(obj) { console.log("Color Set", obj); });
@@ -1487,7 +1484,9 @@ app.model.sisbot = {
 			// this.set('wait_for_send','true');
 			var remember_level = +level;
 			this._update_sisbot('set_led_offset', { offset: remember_level }, function (obj) {
-				// do nothing
+				// save value
+				self.save_to_sisbot(self.get('edit'), null);
+
 				// var end = +new Date();
 				// console.log("Brightness Response (millis):", end-start);
 
@@ -1510,21 +1509,41 @@ app.model.sisbot = {
 		if (this.get('data.installing_updates') == 'true') return this;
 
 		app.plugins.n.notification.confirm('If your RGBW lights are already working, you do not need this step. Continue?', function(resp_num) {
-			if (resp_num < 2) return self;
+			if (resp_num == 1) return self;
 
 			// install_python
 			console.log("Install Python!", resp_num);
 
-			self._update_sisbot('install_python', {}, function(obj) {
-				if (obj.err) {
-					self.set('data.installing_updates_error', 'There was an error updating your Sisbot');
-				} else if (obj.resp) {
-					app.manager.intake_data(obj.resp);
-				}
+			self.save_to_sisbot(self.get('edit'), function(obj) {
+				if (obj.err) return console.log("Save error", obj.err);
+
+				self._update_sisbot('install_python', {}, function(obj) {
+					if (obj.err) {
+						self.set('data.installing_updates_error', 'There was an error updating your Sisbot');
+					} else if (obj.resp) {
+						app.manager.intake_data(obj.resp);
+					}
+				});
 			});
 
 			return this;
 		}, 'Install Python?', ['Cancel', 'OK']);
+	},
+	homing_offset: function(level) {
+		console.log("Homing Offset: ", level);
+		this.set('edit.table_settings.homingOffset',+level);
+	},
+	homing_offset_up: function () {
+		var level = +this.get('edit.table_settings.homingOffset');
+		level += 0.01;
+		if (level > 0.12) level = 0.12;
+		this.set('edit.table_settings.homingOffset',level);
+	},
+	homing_offset_down: function () {
+		var level = +this.get('edit.table_settings.homingOffset');
+		level -= 0.01;
+		if (level < -0.12) level = -0.12;
+		this.set('edit.table_settings.homingOffset',level);
 	},
 	disconnect: function () {
 		app.plugins.n.notification.confirm('Are you sure you want to disconnect from the Sisyphus?', function(resp_num) {
@@ -1716,7 +1735,7 @@ app.model.sisbot = {
 		app.plugins.n.notification.confirm("Changing these settings may cause your table to not function as expected.",
 			function(resp_num) {
 				console.log("Confirm resp", resp_num);
-				if (resp_num !== 2) {
+				if (resp_num == 1) {
 					return self;
 				} else {
 					self.save_to_sisbot(self.get('edit'), function(obj) {
@@ -1731,9 +1750,28 @@ app.model.sisbot = {
 						self.restart();
 					});
 				}
-			}, 'Confirm', ['Save','Cancel']);
+			}, 'Confirm', ['Cancel','Save']);
 	},
 	/******************** VERSIONING ******************************************/
+	check_for_csons: function() {
+		var self = this;
+
+		// update table_settings value
+		this.set('edit.table_settings.cson', this.get('data.cson'));
+		this.set('data.table_settings.cson', this.get('data.cson'));
+
+		// init homingOffset for slider
+		if (this.get('edit.table_settings.homingOffset') == undefined) this.set('edit.table_settings.homingOffset', 0);
+		if (this.get('edit.is_servo') == 'true') this.set('edit.table_settings.homingOffset', 0);
+
+		if (this.get('csons') == 'false') { // only bother loading once
+			this._update_sisbot('get_csons', {}, function(cbb) {
+				console.log("CSONs", cbb.err, cbb.resp);
+
+				self.set('csons', cbb.resp);
+			});
+		}
+	},
 	check_for_version_update: function () {
 		var self	= this;
 		var cbs		= 2;