+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+#2019-10-2 1.8.62
+ - Added multipl download to community and add to playlist. 
+ - Added community model
+ - reworked download_wc()
+#2019-10-2 1.8.60
+ - Added conditional for _processing_registration() to fix re-signup error.
+ - Spinner added when downloading tracks in Community.
+ - Error handeling for email and username in Community updated.
+ - Adjusted thumbnail generation timeout to 9000
+ - Changed stroke_width to 7 in gen_thumbnails.js and model.track
+#2019-09-25 1.8.58
+ - Color Picker Input field, 
+ - Primary/Secondary Buttons incorporate white value
+#2019-09-24 1.8.57
+ - Alert "Track Added to Library" removed.
+ - Save after White slider change
+ - Save after led offset change
+ - Disable Jimmy & Homing offset for servo
+ - Disable auto-brightness in onboarding page
+ - Log files page loads list from sisbot
+ - Paint pattern added
+ - Color picker bugfixes
+ - Color picker input field
+ - Primary/Secondary buttons incorporate white value
+#2019-09-19 1.8.54
+ - CSON selection in Advanced Table Settings
+ - Homing Offset in Advanced Table Settings
+ - Save Advanced Table Settings before installing Python from button
+ - CSON RGBW Offset
+ - Back button fix from Adding Custom Tracks
+#2019-09-12 1.8.50 & 51
+ - Added Add to playlist to the Community when downloading tracks.
+ - Moved sing_in /sign_up functionality to model.session
+ - Added eye icons to show password instead of wording and checkbox.
+ - Remember now stores email and password to session and auto log_in when coming back to Coommunity after closing app.
+ - Removed back arrow from Track Search and Night Mode
 #2019-09-12 1.8.49
  - Install Python added to Advanced Table Settings
  - Demo pattern added to Advanced Lights Settings
@@ -39,7 +75,7 @@
  - Sisbot updates included in regression testing.
  - removed pull-top from Community header to address white gap on Android
  - added new splash screen for ios to address default cordova photo showing. new png is 1125x2436
- - community is "//commeneted out" in this release.
+ - community is "//commented out" in this release.
 #2019-8-15 1.8.32 & 33
  - Community UI changes and re-configuring of pages to flow seemlessly with iPhone gap and notch
  - Added scrolling to all pages in Community.